+++ R/ADeval.r
@@ -23,8 +23,8 @@
 #' @examples
 #' #Simulate 1000 trials with one interim analysis
 #' simdat <- lapply(1:1000, function(i)
-#'  ADsim(n01 = c(100,100), p01 = c(0.4, 0.2), nia = 1))
-#'  simdat <- do.call(rbind, simdat)
+#' 		ADsim(n01 = c(100,100), par01 = c(0.2, 0.4), nia = 1))
+#' simdat <- do.call(rbind, simdat)
 #'
 #' # Define stop for futility stoppin (z<0) and efficacy stopping (z>2)
 #' zcutoff<-matrix(c(0,2),1,2)
@@ -40,8 +40,8 @@ ADeval<-function(simdat, zcutoff, finalm = c("z","ci"), alpha = 0.025) {
 	#derive n and true values from simdat
 	nia<-sum(grepl("ia[0-9]_pe",colnames(simdat)))
 	n<-mean(simdat[,"fa_n0"]) + mean(simdat[,"fa_n1"])
-	tp0<-mean(simdat[,"true_p0"])
-	tp1<-mean(simdat[,"true_p1"])
+	#tp0<-mean(simdat[,"true_p0"])
+	#tp1<-mean(simdat[,"true_p1"])
 	rdt<-mean(simdat[,"true_pe"])
 
 	#derive direction from pe and z
@@ -56,13 +56,14 @@ ADeval<-function(simdat, zcutoff, finalm = c("z","ci"), alpha = 0.025) {
 
 	resi<-data.frame(simdat)
 
-	resi$obs_p0<-resi$fa_x0/resi$fa_n0
-	resi$obs_p1<-resi$fa_x1/resi$fa_n1
+	fans<-colnames(simdat)[grepl("^fa_",colnames(simdat))]
+	fan<-fans[1]
+	for (fan in fans) {
+		nname<-sub("fa_","obs_",fan)
+		resi<-cbind(resi,nname=resi[,fan])
+		colnames(resi)[colnames(resi)=="nname"]<-nname
+	}
 
-	resi$obs_pe<-resi$fa_pe
-	resi$obs_lci<-resi$fa_lci
-	resi$obs_uci<-resi$fa_uci
-	resi$obs_z<-resi$fa_z
 
 	if (nia>0) {
 		sel<-data.frame(matrix(NA,nrow(resi),nia))
@@ -83,23 +84,17 @@ ADeval<-function(simdat, zcutoff, finalm = c("z","ci"), alpha = 0.025) {
 
 		#use rd at interim at the first stopped rd
 		resi$tstop<-apply(sel,1,function(x) ifelse(any(x>0),min(which(x>0)),NA))
-		resi$neff<-n
+		#resi$neff<-n
 
 		i<-1
 		for (i in (1:nia)) {
-		  si<-!is.na(resi$tstop) & resi$tstop==i
-
-		  resi$obs_p0[si]<-resi[si,paste0("ia",i,"_x0")]/resi[si,paste0("ia",i,"_n0")]
-		  resi$obs_p1[si]<-resi[si,paste0("ia",i,"_x1")]/resi[si,paste0("ia",i,"_n1")]
-
-		  resi$obs_pe[si]<-resi[si,paste0("ia",i,"_pe")]
-		  resi$obs_lci[si]<-resi[si,paste0("ia",i,"_lci")]
-		  resi$obs_uci[si]<-resi[si,paste0("ia",i,"_uci")]
-
-		  niobs<-mean(resi[,paste0("ia",i,"_n0")] + resi[,paste0("ia",i,"_n1")])
-		  resi$neff[si]<-niobs
-
-		  resi$obs_z[si]<-resi[si,paste0("ia",i,"_z")]
+			si<-!is.na(resi$tstop) & resi$tstop==i
+			
+		  	for (fan in fans) {
+				nname<-sub("fa_","obs_",fan)
+				iname<-sub("fa_",paste0("ia",i,"_"),fan)
+				resi[si,nname]<-resi[si,iname]
+			}		
 		}
 
 		resi$tstop_fut<-apply(sel,1,function(x) ifelse(any(x==1),min(which(x==1)),NA))
@@ -113,20 +108,23 @@ ADeval<-function(simdat, zcutoff, finalm = c("z","ci"), alpha = 0.025) {
 
 	} else {
 		resi$anystop<-"Not stopped"
-		niobs<-mean(resi[,"fa_n0"] + resi[,"fa_n1"])
-		resi$neff<-niobs
+
 	}
 
 
 	resi$anystop<-factor(resi$anystop,
 		levels=c("Not stopped","Stopped for futility","Stopped for efficacy"))
-
-	av_p0<-mean(resi$obs_p0)
-	av_p1<-mean(resi$obs_p1)
-	av_pe<-mean(resi$obs_pe)
-	av_lci<-mean(resi$obs_lci)
-	av_uci<-mean(resi$obs_uci)
-
+	
+	
+	#average summarize: 
+	avs<-numeric(0)
+	for (fan in fans) {
+		oname<-sub("fa_","obs_",fan)
+		avs<-append(avs,mean(resi[,oname]))	
+	}
+	names(avs)<-gsub("fa_","av_",fans)
+	
+	#power 
 	if (finalm == "ci") {
 		if (direct=="lower") {
 			pwr<-mean(resi$obs_uci<0)
@@ -151,24 +149,26 @@ ADeval<-function(simdat, zcutoff, finalm = c("z","ci"), alpha = 0.025) {
 			resi$anystop %in% c("Stopped for futility","Stopped for efficacy")))
 	}
 	nastopi<-c(paste0("pstop_ia",1:nia),paste0("pstop_fut_ia",1:nia),paste0("pstop_eff_ia",1:nia))
-
-	avn<-mean(resi$neff)
-	minn<-min(resi$neff)
-	maxn<-max(resi$neff)
+	
+	#expected total ss 
+	av_n<-avs["av_n0"] + avs["av_n1"] 
+	min_n<-min(resi$obs_n0 + resi$obs_n1)
+	max_n<-max(resi$obs_n0 + resi$obs_n1)
 	bias<-mean(resi$obs_pe - rdt)
 	sd<-sd(resi$obs_pe)
 	mse<-mean((resi$obs_pe - rdt)^2)
 	coverage<-mean(resi$obs_lci<rdt & resi$obs_uci>rdt)
 
-	op<-c(tp0,tp1,rdt,
-		av_p0,av_p1,av_pe,av_lci,av_uci,
-		pwr,pstop,pstopfut,pstopeff,pstopi,pstopfuti,pstopeffi,
-			avn,minn,maxn,bias,sd,mse,coverage)
+	op<-c(rdt,avs,
+		pwr,pstop,pstopfut,pstopeff,
+			pstopi,pstopfuti,pstopeffi,
+			av_n,min_n,max_n,
+			bias,sd,mse,coverage)
 
-	names(op)<-c("true_p0","true_p1","true_pe",
-		"av_p0","av_p1","av_pe","av_lci","av_uci",
+	names(op)<-c("true_pe",names(avs),
 		"psig","pstop","pstop_fut","pstop_eff",nastopi,
-		"avn","minn","maxn","bias","sd","mse","coverage")
+		"av_n","min_n","max_n",
+		"bias","sd","mse","coverage")
 
 	res<-list(resi,op)
 	names(res)<-c("rawdata","opchar")