+++ NEWS.md
@@ -15,3 +15,8 @@ AdaptiveDesignSim 0.1.0
 ------------------
 Renaming and major update of the functions
 Added possiblity for user-define estimation functions
+
+AdaptiveDesignSim 0.2.0
+------------------
+Add simfun for flexible simulation
+Add default function for continuous variable